Disable rooting restriction for Samsung devices
This is mainly useful for the recovery. In the default Samsung kernel,
there is a restriction that disables running a process named 'adbd' as
root. Simply renaming it to e.g. 'adbe' fixes it, but this disables this
restriction (at least for the recovery).

Thanks to Dees_Troy from #twrp on Freenode.

Change-Id: I8573bb0da6cdbfc7af083b31e13c818353cf7e58